+++ src/options.jl
@@ -121,16 +121,16 @@ Base.@kwdef mutable struct CcOptions
   """`⟨false⟩` decompose full doubles amplitudes in SVD-DCSD (slow). """
   decompose_full_doubles::Bool = false
   """`⟨"cc_amplitudes"⟩` main part of filename for start amplitudes. 
-      For example, the singles amplitudes are read from `start*"_singles"` """
+      For example, the singles amplitudes are read from `start*"_1"` """
   start::String = "cc_amplitudes"
   """`⟨"cc_amplitudes"⟩` main part of filename to save amplitudes.
-      For example, the singles amplitudes are saved to `save*"_singles"` """
+      For example, the singles amplitudes are saved to `save*"_1"` """
   save::String = "cc_amplitudes"
   """`⟨"cc_multipliers"⟩` main part of filename for start Lagrange multipliers. 
-      For example, the singles Lagrange multipliers are read from `start_lm*"_singles"` """
+      For example, the singles Lagrange multipliers are read from `start_lm*"_1"` """
   start_lm::String = "cc_multipliers"
   """`⟨"cc_multipliers"⟩` main part of filename to save Lagrange multipliers.
-      For example, the singles Lagrange multipliers are saved to `save_lm*"_singles"` """
+      For example, the singles Lagrange multipliers are saved to `save_lm*"_1"` """
   save_lm::String = "cc_multipliers"
   """`⟨0⟩` Don't use MP2 amplitudes as starting guess for the CC amplitudes """
   nomp2::Int = 0